python/requirements.txt
# =============================================================================
# Accounting Automation System - Python Dependencies
# =============================================================================
# Install: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Core
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
alembic>=1.13.0

# -----------------------------------------------------------------------------
# Claude API (Anthropic)
# -----------------------------------------------------------------------------
anthropic>=0.18.0

# -----------------------------------------------------------------------------
# Excel & PowerPoint Generation
# -----------------------------------------------------------------------------
openpyxl>=3.1.2
python-pptx>=0.6.23
xlsxwriter>=3.1.9

# -----------------------------------------------------------------------------
# PDF Generation & Processing
# -----------------------------------------------------------------------------
reportlab>=4.0.8
pypdf>=3.17.0
pdf2image>=1.17.0  # For PDF to image conversion (Claude Vision)
pillow>=10.1.0

# -----------------------------------------------------------------------------
# Data Processing
# -----------------------------------------------------------------------------
pandas>=2.1.4
numpy>=1.26.2

# -----------------------------------------------------------------------------
# HTTP & API Clients
# -----------------------------------------------------------------------------
httpx>=0.26.0
requests>=2.31.0
aiohttp>=3.9.1

# -----------------------------------------------------------------------------
# QuickBooks API
# -----------------------------------------------------------------------------
intuit-oauth>=1.2.6
python-quickbooks>=0.9.4

# -----------------------------------------------------------------------------
# Google APIs
# -----------------------------------------------------------------------------
google-api-python-client>=2.111.0
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
gspread>=6.0.0

# -----------------------------------------------------------------------------
# Telegram Bot
# -----------------------------------------------------------------------------
python-telegram-bot>=20.7

# -----------------------------------------------------------------------------
# YAML & Config
# -----------------------------------------------------------------------------
pyyaml>=6.0.1
jinja2>=3.1.2

# -----------------------------------------------------------------------------
# Date & Time
# -----------------------------------------------------------------------------
python-dateutil>=2.8.2
pytz>=2023.3

# -----------------------------------------------------------------------------
# Validation
# -----------------------------------------------------------------------------
email-validator>=2.1.0
phonenumbers>=8.13.27

# -----------------------------------------------------------------------------
# Charts & Visualization
# -----------------------------------------------------------------------------
matplotlib>=3.8.2
seaborn>=0.13.0

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest>=7.4.3
pytest-asyncio>=0.23.2
pytest-cov>=4.1.0
pytest-mock>=3.12.0
faker>=21.0.0
responses>=0.24.1

# -----------------------------------------------------------------------------
# Code Quality
# -----------------------------------------------------------------------------
black>=23.12.1
isort>=5.13.2
flake8>=6.1.0
mypy>=1.7.1
types-requests>=2.31.0.10
types-PyYAML>=6.0.12.12

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
tenacity>=8.2.3  # Retry logic
structlog>=23.2.0  # Structured logging
cryptography>=41.0.7  # For encryption
